% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dataset.R
\name{clean.dataset}
\alias{clean.dataset}
\title{Dataset Cleansing}
\usage{
clean.dataset(data, clean.nan = TRUE, clean.ohe = 10)
}
\arguments{
\item{data}{\code{[n, d]} array/matrix with the \code{n} samples in \code{d} dimensions.}

\item{clean.nan}{whether to remove samples with invalid entries. Defaults to \code{TRUE}.
\itemize{
\item{\code{TRUE}}{Remove samples that have features with \code{NaN} or non-finite.}
\item{\code{FALSE}}{Do not remove samples.}
}}

\item{clean.ohe}{options for whether to one-hot-encode columns. Defaults to \code{10}.
\itemize{
\item{\code{clean.ohe < 1}}{Converts columns with < thr*n unique identifiers to one-hot encoded.}
\item{\code{is.integer(clean.ohe)}}{Converts columns with < thr unique identifiers to one-hot encoded.}
\item{\code{FALSE}}{Do not one-hot-encode any columns.}
}}
}
\value{
A list containing the following:
\itemize{
\item{X}{\code{[m, d+r]} the array with \code{m} samples in \code{d+r} dimensions, where \code{r} is the number of additional columns appended for encodings. \code{m < n} when  there are non-finite entries. \code{colnames(data)} returns the column names of the cleaned columns.}
\item{samples}{\code{m} the sample ids that are included in the final array, where \code{samp[i]} is the original row id corresponding to \code{Xc[i,]}. If \code{m < n}, there were non-finite entries that were purged.}
}
}
\description{
A function for scrubbing a dataset for usage with most standard algorithms. This involves one-hot-encoding columns that are probably categorical.
}
\author{
Eric Bridgeford
}
