% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmlb.R
\name{pmlb.list}
\alias{pmlb.list}
\title{List from PMLB Dataset}
\usage{
pmlb.list(datasets = NULL, tasks = NULL)
}
\arguments{
\item{tasks}{the type of the task, either "classification" or "regression". Defaults to \code{NULL}.
\itemize{
\item{\code{NULL}}{Return all datasets matching the desired query.}
\item{\code{'classification'}}{Load all classification datasets matching the desired query.}
\item{\code{'regression'}}{Load all regression datasets matching the desired query.}
}}

\item{dataset}{the name of the dataset you wish to load. Defaults to \code{NULL}.
\itemize{
\item{\code{NULL}}{Load all the datasets without specifying a specific name matching the desired query.}
\item{\code{'datasetid'}}{Returns the dataset with the desired id \code{'datasetid'} matching the desired query.}
}}
}
\value{
A list containing the following:
\itemize{
\item{\code{dsets.meta}}{A \code{[length(tasks)]} element list, where each sublist contains a \code{tibble} where each row contains relevant metadata for the desired dataset.}
\item{\code{dsets.info}}{A \code{[n]} list where each element is named after a respective dataset, and each element contains the \code{did} and \code{task} of that dataset.}
}
}
\description{
A function to list available datasets for a specific task from the PMLB dataset. Also returns useful metadata.
}
\examples{
require(slbR)

# get information about all of the classification tasks in pmlb
test <- pmlb.list(tasks="classification")

# get information about a specific dataset in pmlb
test <- pmlb.list(datasets="adult")

# get information about 2 specific datasets in pmlb
test <- pmlb.list(datasets=c("adult", "chscase_geyser1"))
}
\author{
Eric Bridgeford
}
