% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{slb.load.datasets}
\alias{slb.load.datasets}
\title{Load Datasets from Available Repositories}
\usage{
slb.load.datasets(repositories = NULL, datasets = NULL, tasks = NULL,
  clean.invalid = TRUE, clean.ohe = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{repositories}{the name of the repository you would like to query a dataset from. Defaults to \code{NULL}, which will return datasets matching from all repositories.
\itemize{
\item \code{NULL} Load datasets from all repositories matching other queries.
\item \code{"pmlb"} Load datasets from the Penn Machine-Learning Benchmarks.
\item \code{"uci"} Load datasets from the University of California - Irvine Machine Learning Repository.
\item \code{"mnist"} Load datasets from the MNIST dataset.
\item \code{c("repo1", "repo2", ...)} Load data from the indicated repositories.
}}

\item{datasets}{the name of the dataset you wish to load. Defaults to \code{NULL}.
\itemize{
\item \code{NULL} Load all the datasets without specifying a specific name matching the desired query.
\item \code{'datasetid'} Returns the dataset with the desired id matching the desired query.
\item \code{c("datasetid1", "datasetid2", ...)} Load data from the indicated datasets.
}}

\item{tasks}{the type of the task, either "classification" or "regression". Defaults to \code{NULL}.
\itemize{
\item \code{NULL} Return all datasets matching the desired query.
\item \code{'classification'} Load all classification datasets matching the desired query.
\item \code{'regression'} Load all regression datasets matching the desired query.
\item \code{c("taskid1", "taskid2", ...)} Load data for the indicated tasks.
}}

\item{clean.invalid}{whether to remove samples with invalid entries. Defaults to \code{TRUE}.
\itemize{
\item \code{TRUE} Remove samples that have features with \code{NaN} entries or non-finite.
\item \code{FALSE} Do not remove samples that have features with \code{NaN} entries or are non-finite..
}}

\item{clean.ohe}{options for whether to one-hot-encode columns. Defaults to \code{FALSE}.
\itemize{
\item \code{clean.ohe < 1} Converts columns with < thr*n unique identifiers to one-hot encoded.
\item \code{is.integer(clean.ohe)} Converts columns with < thr unique identifiers to one-hot encoded.
\item \code{FALSE} Do not one-hot-encode any columns.
}}

\item{verbose}{whether to print messages to the console if a repository or dataset is being ignored. defaults to \code{TRUE}.}

\item{...}{trailing args.}
}
\value{
A list of lists, where each element is a key-worded list for a particular benchmark dataset, containing at least the following:
\item{\code{X}}{\code{[n, d]} array with the \code{n} samples in \code{d} dimensions.}
\item{\code{Y}}{\code{[n]} vector or \code{[n, r]} array with responses for each of the \code{n} samples.}
}
\description{
A function to load a specified dataset from the PMLB dataset.
}
\examples{
library(slb)
# request 1 specific dataset from the pmlb dataset
test <- slb.load.datasets(repositories="pmlb", datasets="ESL", clean.invalid=FALSE, clean.ohe=FALSE)
length(test$ESL$Y) == 488 # a known example from the pmlb dataset

# request all of the pmlb classification datasets
\dontrun{
test <- slb.load.datasets(repositories="pmlb", tasks="classification")
length(test) <- 166  # validates that we loaded all of the classification datasets from pmlb
}
}
\author{
Eric Bridgeford
}
