% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmlb.R
\name{pmlb.load}
\alias{pmlb.load}
\title{Load from PMLB Dataset}
\usage{
pmlb.load(datasets = NULL, tasks = NULL)
}
\arguments{
\item{tasks}{the type of the tasks to return datasets for, either "classification" or "regression". Defaults to \code{NULL}.
\itemize{
\item{\code{NULL}}{Return all datasets matching the desired query.}
\item{\code{'classification'}}{Load all classification datasets matching the desired query.}
\item{\code{'regression'}}{Load all regression datasets matching the desired query.}
}}

\item{dataset}{the name of the dataset you wish to load. Defaults to \code{NULL}.
\itemize{
\item{\code{NULL}}{Load all the datasets without specifying a specific name matching the desired query.}
\item{\code{'datasetid'}}{Returns the dataset with the desired id \code{'datasetid'} matching the desired query.}
}}

\item{clean}{whether to clean the dataset, using \link{dataset.clean}.}
}
\value{
A list containing the following:
\itemize{
\item{\code{data}}{The data for each dataset, as a list of the following:}
\itemize{
\item{X}{\code{[n, d]} array with the \code{n} samples in \code{d} dimensions.}
\item{Y}{\code{[n]} vector with labels for each of the \code{n} samples. This item is only returned in the event that the dataset is a classification task.}
}
\item{\code{dsets.info}}{Useful metadata for each dataset. See \link{pmlb.list} for details.}
}
}
\description{
A function to load a specified dataset from the PMLB dataset.
}
\examples{
require(slbR)

# load a specific dataset from pmlb
test <- pmlb.load(datasets="adult")
length(test$data) == 1  # the number of datasets requested
length(test$data$adult$Y) == 48842  # the number of known features in pmlb

test <- pmlb.load(tasks="classification")
length(test$data) == 166  # the number of classification tasks in pmlb

# load 2 datasets from pmlb
test <- pmlb.load(datasets=c("adult", "chscase_geyser1"))
length(test$data) == 2
}
\author{
Eric Bridgeford
}
